chore: Header 컴포넌트 인터페이스 수정
