--- app/layout.css.ts
@@ -1,7 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-import { vars } from '@/src/styles/theme.css';
-
-export const logo = style({
-  color: vars.colors.primary,
-});
