+++ src/components/layout/Header/index.tsx
@@ -1,18 +1,20 @@
-import { PropsWithChildren } from 'react';
+import Link from 'next/link';
 
 import * as styles from '@/src/components/layout/Header/index.css';
 
-export default function Header({ children }: PropsWithChildren) {
-  return <header className={styles.header}>{children}</header>;
+export default function Header() {
+  return (
+    <header className={styles.header}>
+      <div className={styles.headerLeft}>
+        <Link className={styles.logo} href="/">
+          TECH
+        </Link>
+      </div>
+      <div className={styles.headerRight}>
+        <Link href="/">Home</Link>
+        <Link href="/login">Sign in</Link>
+        <Link href="/register">Sign up</Link>
+      </div>
+    </header>
+  );
 }
-
-function Left({ children }: PropsWithChildren) {
-  return <div className={styles.headerLeft}>{children}</div>;
-}
-
-function Right({ children }: PropsWithChildren) {
-  return <div className={styles.headerRight}>{children}</div>;
-}
-
-Header.Left = Left;
-Header.Right = Right;
